Move ErrorLabelComponent and ErrorMessageGenerator to error folder inside MessagesArea
