+++ src/panel.module.ts
@@ -10,11 +10,11 @@ import { LoginComponent } from "./login.component";
 import { RegisterComponent } from "./register.component";
 import { SidebarComponent } from "./sidebar.component";
 import { SidebarItemsComponent } from "./sidebar-items.component";
-import { ErrorMessageComponent } from "./errors-area/error-message.component";
-import { ErrorsAreaComponent } from "./errors-area/errors-area.component";
+import { MessageComponent } from "./messages-area/message.component";
+import { MessagesAreaComponent } from "./messages-area/messages-area2.component";
 import { MenuBarComponent } from "./menu-bar.component";
 import { PaginatorComponent } from "./paginator/paginator.component";
-import { ErrorLabelComponent } from "./errors-area/error-label.component";
+import { ErrorLabelComponent } from "./messages-area/error/error-label.component";
 
 // Modules
 import { SemanticModule } from "./semantic/semantic.module";
@@ -25,7 +25,7 @@ import { RouterService } from "./router.service";
 import { HeaderService } from "./header.service";
 import { SidebarService } from "./sidebar.service";
 import { MyAppsSidebarService } from "./my-apps/my-apps-sidebar.service";
-import { ErrorsAreaService } from "./errors-area/errors-area.service";
+import { MessagesAreaService } from "./messages-area/messages-area.service";
 
 @NgModule( {
 	imports: [
@@ -42,8 +42,8 @@ import { ErrorsAreaService } from "./errors-area/errors-area.service";
 		RegisterComponent,
 		SidebarComponent,
 		SidebarItemsComponent,
-		ErrorMessageComponent,
-		ErrorsAreaComponent,
+		MessageComponent,
+		MessagesAreaComponent,
 		MenuBarComponent,
 		PaginatorComponent,
 		ErrorLabelComponent,
@@ -54,8 +54,8 @@ import { ErrorsAreaService } from "./errors-area/errors-area.service";
 		RegisterComponent,
 		SidebarComponent,
 		MenuBarComponent,
-		ErrorMessageComponent,
-		ErrorsAreaComponent,
+		MessageComponent,
+		MessagesAreaComponent,
 		PaginatorComponent,
 		ErrorLabelComponent,
 	],
@@ -67,7 +67,7 @@ export class PanelModule {
 	static forRoot():ModuleWithProviders {
 		return {
 			ngModule: PanelModule,
-			providers: [ HeaderService, SidebarService, RouterService, MyAppsSidebarService, ErrorsAreaService ]
+			providers: [ HeaderService, SidebarService, RouterService, MyAppsSidebarService, MessagesAreaService ]
 		};
 	}
 }
